requirements.txt
torch>=2.0.0
triton>=2.1.0
pyyaml>=6.0
pytest>=7.0
numpy>=1.24.0
